requirements.txt
fastapi
numpy
uvicorn
langchain
langchain-community
langchain-openai
python-dotenv
chromadb
PyPDF2
pydantic